+++ server/camcops_server/tasks/chit.py
@@ -29,16 +29,22 @@ camcops_server/tasks/chit.py
 
 """
 
+from typing import List, Type, Tuple, Dict, Any
+
+from cardinal_pythonlib.classes import classproperty
+from cardinal_pythonlib.stringfunc import strseq
+from semantic_version import Version
+from sqlalchemy import Integer
+from sqlalchemy.ext.declarative import DeclarativeMeta
+
 from camcops_server.cc_modules.cc_constants import CssClass
 from camcops_server.cc_modules.cc_db import add_multiple_columns
 from camcops_server.cc_modules.cc_html import (
     tr_qa,
-    get_yes_no_unknown,
     tr,
     answer,
 )
 from camcops_server.cc_modules.cc_request import CamcopsRequest
-from camcops_server.cc_modules.cc_sqla_coltypes import BoolColumn
 from camcops_server.cc_modules.cc_summaryelement import SummaryElement
 from camcops_server.cc_modules.cc_task import (
     TaskHasPatientMixin,
@@ -46,10 +52,6 @@ from camcops_server.cc_modules.cc_task import (
     get_from_dict,
 )
 from camcops_server.cc_modules.cc_text import SS
-from cardinal_pythonlib.stringfunc import strseq
-from sqlalchemy import Integer
-from sqlalchemy.ext.declarative import DeclarativeMeta
-from typing import List, Type, Tuple, Dict, Any
 
 
 class ChitMetaclass(DeclarativeMeta):
@@ -65,9 +67,9 @@ class ChitMetaclass(DeclarativeMeta):
             "q",
             1,
             cls.N_SCORED_QUESTIONS,
-            minimum=0,
-            maximum=3,
-            comment_fmt="Q{n} ({s}) (0 strongly disagree - 3 strongly agree)",
+            minimum=cls.MIN_ANSWER,
+            maximum=cls.MAX_ANSWER,
+            comment_fmt="Q{n} ({s}) (0 strongly disagree - 4 strongly agree)",
             comment_strings=[
                 "hate unfinished task",
                 "just right",
@@ -87,12 +89,6 @@ class ChitMetaclass(DeclarativeMeta):
             ],
         )
 
-        setattr(
-            cls,
-            "q16",
-            BoolColumn("q16", comment="Q16 (negative effect) (0 no, 1 yes)"),
-        )
-
         super().__init__(name, bases, classdict)
 
 
@@ -101,15 +97,20 @@ class Chit(TaskHasPatientMixin, Task, metaclass=ChitMetaclass):
     shortname = "CHI-T"
 
     N_SCORED_QUESTIONS = 15
-    N_QUESTIONS = 16
-    MAX_SCORE_MAIN = 3 * N_SCORED_QUESTIONS
+    MIN_ANSWER = 0
+    MAX_ANSWER = 4
+    MAX_SCORE_MAIN = MAX_ANSWER * N_SCORED_QUESTIONS
     SCORED_QUESTIONS = strseq("q", 1, N_SCORED_QUESTIONS)
-    ALL_QUESTIONS = strseq("q", 1, N_QUESTIONS)
 
     @staticmethod
     def longname(req: "CamcopsRequest") -> str:
         _ = req.gettext
         return _("Cambridge–Chicago Compulsivity Trait Scale")
+
+    # noinspection PyMethodParameters
+    @classproperty
+    def minimum_client_version(cls) -> Version:
+        return Version("2.4.15")
 
     def get_summaries(self, req: CamcopsRequest) -> List[SummaryElement]:
         return self.standard_task_summary_fields() + [
@@ -122,7 +123,7 @@ class Chit(TaskHasPatientMixin, Task, metaclass=ChitMetaclass):
         ]
 
     def is_complete(self) -> bool:
-        if self.any_fields_none(self.ALL_QUESTIONS):
+        if self.any_fields_none(self.SCORED_QUESTIONS):
             return False
         if not self.field_contents_valid():
             return False
@@ -134,11 +135,10 @@ class Chit(TaskHasPatientMixin, Task, metaclass=ChitMetaclass):
     def get_task_html(self, req: CamcopsRequest) -> str:
         score_dict = {
             None: None,
-            0: "0 — " + self.wxstring(req, "a0"),
-            1: "1 — " + self.wxstring(req, "a1"),
-            2: "2 — " + self.wxstring(req, "a2"),
-            3: "3 — " + self.wxstring(req, "a3"),
         }
+
+        for i in range(self.MIN_ANSWER, self.MAX_ANSWER + 1):
+            score_dict[i] = f"{i} — " + self.wxstring(req, f"a{i}")
 
         rows = ""
         for i in range(1, self.N_SCORED_QUESTIONS + 1):
@@ -147,10 +147,6 @@ class Chit(TaskHasPatientMixin, Task, metaclass=ChitMetaclass):
             answer_cell = get_from_dict(score_dict, getattr(self, q_field))
 
             rows += tr_qa(question_cell, answer_cell)
-
-        rows += tr_qa(
-            "16. " + self.wxstring(req, "q16"), get_yes_no_unknown(req, "q16")
-        )
 
         html = """
             <div class="{CssClass.SUMMARY}">
@@ -162,12 +158,16 @@ class Chit(TaskHasPatientMixin, Task, metaclass=ChitMetaclass):
             <table class="{CssClass.TASKDETAIL}">
                 <tr>
                     <th width="60%">Question</th>
-                    <th width="40%">Answer</th>
+                    <th width="40%">Answer <sup>[2]</sup></th>
                 </tr>
                 {rows}
             </table>
             <div class="{CssClass.FOOTNOTES}">
-                [1] Sum for questions 1–15.
+                [1] Sum for questions 1–15. Prior to CamCOPS version 2.4.15
+                each question scored 0–3 with a maximum possible score of 45.
+                [2] Prior to CamCOPS version 2.4.15 the responses were:
+                0 — Strongly disagree, 1 — Disagree, 2 — Agree, 3 — Strongly
+                agree.
             </div>
         """.format(
             CssClass=CssClass,