+++ tablet_qt/menu/researchmenu.cpp
@@ -30,7 +30,7 @@
 #include "tasks/chit.h"
 #include "tasks/diagnosisicd9cm.h"
 #include "tasks/ided3d.h"
-#include "tasks/isaaq.h"
+#include "tasks/isaaq10.h"
 #include "tasks/isaaqed.h"
 #include "tasks/kirby.h"
 #include "tasks/mfi20.h"
@@ -77,7 +77,7 @@ void ResearchMenu::makeItems()
         ),
         MAKE_TASK_MENU_ITEM(CardinalExpDetThreshold::CARDINALEXPDETTHRESHOLD_TABLENAME, m_app),
         MAKE_TASK_MENU_ITEM(CardinalExpectationDetection::CARDINALEXPDET_TABLENAME, m_app),
-        MAKE_TASK_MENU_ITEM(Isaaq::ISAAQ_TABLENAME, m_app),
+        MAKE_TASK_MENU_ITEM(Isaaq10::ISAAQ10_TABLENAME, m_app),
         MAKE_TASK_MENU_ITEM(IsaaqEd::ISAAQED_TABLENAME, m_app),
     };
 }