+++ tablet_qt/tools/build_qt.py
@@ -974,6 +974,15 @@ class Platform(object):
         return self.desktop or self.ios
 
     # -------------------------------------------------------------------------
+    # Build Qt with OpenSSL or use native SSL library
+    # -------------------------------------------------------------------------
+    @property
+    def use_openssl_with_qt(self) -> bool:
+        # Although we use SecureTransport instead of OpenSSL on iOS, we still
+        # need OpenSSL for SQLCipher
+        return not self.ios
+
+    # -------------------------------------------------------------------------
     # Library (e.g. .so, DLL) verification
     # -------------------------------------------------------------------------
 
@@ -3255,10 +3264,6 @@ def build_qt(cfg: Config, target_platform: Platform) -> str:
     # Qt: Setup
     # -------------------------------------------------------------------------
 
-    # Means by which Qt links to OpenSSL?
-    qt_openssl_linkage_static = (
-        cfg.qt_openssl_static and target_platform.qt_linkage_static
-    )
     # If Qt is linked dynamically, we do not let it link to OpenSSL
     # statically (it won't work).
 
@@ -3273,12 +3278,6 @@ def build_qt(cfg: Config, target_platform: Platform) -> str:
         # right automatically? Yes.
         # if USE_CLANG_NOT_GCC_FOR_ANDROID_ARM:
         #     require(CLANG)
-
-    opensslrootdir, opensslworkdir = cfg.get_openssl_rootdir_workdir(
-        target_platform
-    )
-    openssl_include_root = join(opensslworkdir, "include")
-    openssl_lib_root = opensslworkdir
 
     builddir = cfg.qt_build_dir(target_platform)
     installdir = cfg.qt_install_dir(target_platform)
@@ -3304,6 +3303,32 @@ def build_qt(cfg: Config, target_platform: Platform) -> str:
     # Qt: Environment
     # -------------------------------------------------------------------------
     env = cfg.get_starting_env()
+    crypto_dependencies = ""
+    if target_platform.linux:
+        crypto_dependencies = "-ldl -lpthread"
+    elif target_platform.windows:
+        # Copying openssl/Configurations/10-main.conf
+        crypto_dependencies = "-lws2_32 -lgdi32 -ladvapi32 -lcrypt32 -luser32"
+
+    if target_platform.use_openssl_with_qt:
+        opensslrootdir, opensslworkdir = cfg.get_openssl_rootdir_workdir(
+            target_platform
+        )
+        openssl_include_root = join(opensslworkdir, "include")
+        openssl_lib_root = opensslworkdir
+
+        openssl_libs = (
+            f"-L{openssl_lib_root} -lssl -lcrypto {crypto_dependencies}"
+        )
+
+        # See also https://bugreports.qt.io/browse/QTBUG-62016
+        env["OPENSSL_LIBS"] = openssl_libs
+        # Setting OPENSSL_LIBS as an *environment variable* may be unnecessary,
+        # but is suggested by Qt; http://doc.qt.io/qt-4.8/ssl.html
+        # However, it seems necessary to set it as an *option* to configure;
+        # see below.
+
+    cfg.set_compile_env(env, target_platform)
 
     # -------------------------------------------------------------------------
     # Qt: Directories
@@ -3323,17 +3348,31 @@ def build_qt(cfg: Config, target_platform: Platform) -> str:
     else:
         configure_prog_name = "configure"
     # sysroot = cfg.sysroot(target_platform, env)
-    includedirs = [openssl_include_root]  # #include files for OpenSSL
+    includedirs = []
     objdirs = []  # type: List[str]
-    libdirs = [openssl_lib_root]  # libraries for OpenSSL
+    libdirs = []
     qt_config_cmake_args = []
     # qt_config_cmake_args = ["-DCMAKE_MESSAGE_LOG_LEVEL=STATUS"]  # Verbose?
+
+    if target_platform.use_openssl_with_qt:
+        includedirs.append(openssl_include_root)  # #include files for OpenSSL
+        libdirs.append(openssl_lib_root)  # libraries for OpenSSL
+
     qt_config_args = [
         join(cfg.qt_src_gitdir, configure_prog_name),
         # General options:
         "-prefix",
         installdir.replace("\\", "\\\\"),  # where to install Qt
+        "-recheck-all",  # don't cache from previous configure runs
+        # "-sysroot": not required; Qt's configure should handle this
+        # "-gcc-sysroot": not required
     ]
+
+    if target_platform.use_openssl_with_qt:
+        qt_config_args.append("OPENSSL_LIBS=" + openssl_libs)
+    else:
+        qt_config_args.append("-no-openssl")
+
     if target_platform.qt_linkage_static:
         qt_config_args.append("-static")
         # makes a static Qt library (cf. default of "-shared")
@@ -3424,7 +3463,8 @@ def build_qt(cfg: Config, target_platform: Platform) -> str:
     elif target_platform.ios:
         # http://doc.qt.io/qt-5/building-from-source-ios.html
         # "A default build builds both the simulator and device libraries."
-        qt_config_args += ["-xplatform", "macx-ios-clang"]
+        # Use Apple's own SSL implementation
+        qt_config_args += ["-securetransport", "-xplatform", "macx-ios-clang"]
 
         # Don't build for both simulator and device (the default) as it causes
         # problems with our OpenSSL setup
@@ -3486,20 +3526,24 @@ def build_qt(cfg: Config, target_platform: Platform) -> str:
         # ... release is default in 5.7 (as per "configure -h")
         # ... check with "readelf --debug-dump=decodedline <LIBRARY.so>"
         # ... http://stackoverflow.com/questions/1999654
-        # ... https://forum.qt.io/topic/75056/configuring-qt-what-replaces-debug-and-release/7  # noqa
+        # ... https://forum.qt.io/topic/75056/configuring-qt-what-replaces-debug-and-release/7  # noqa: E501
 
-    # OpenSSL linkage?
-    # For testing a new OpenSSL build, have cfg.qt_openssl_static=False, or you
-    # have to rebuild Qt every time... extremely slow.
-    if qt_openssl_linkage_static:
-        # https://doc-snapshots.qt.io/qt6-6.2/configure-options.html
-        qt_config_args.append("-openssl-linked")  # OpenSSL
-        qt_config_cmake_args.append("-DOPENSSL_USE_STATIC_LIBS=ON")
-    else:
-        qt_config_args += ["-openssl", "yes"]  # OpenSSL
+    if target_platform.use_openssl_with_qt:
+        # OpenSSL linkage?
+        # For testing a new OpenSSL build, have cfg.qt_openssl_static=False, or
+        # you have to rebuild Qt every time... extremely slow.
+        # Means by which Qt links to OpenSSL?
+        qt_openssl_linkage_static = (
+            cfg.qt_openssl_static and target_platform.qt_linkage_static
+        )
+        if qt_openssl_linkage_static:
+            qt_config_args.append("-openssl-linked")  # OpenSSL
+            qt_config_cmake_args.append("-DOPENSSL_USE_STATIC_LIBS=ON")
+        else:
+            qt_config_args += ["-openssl", "yes"]  # OpenSSL
 
-    # Qt's idea of "root" different to our own
-    qt_config_cmake_args.append(f"-DOPENSSL_ROOT_DIR={opensslworkdir}")
+        # Qt's idea of "root" different to our own
+        qt_config_cmake_args.append(f"-DOPENSSL_ROOT_DIR={opensslworkdir}")
 
     if cfg.verbose >= 1:
         # Qt by default sets CMAKE_MESSAGE_LOG_LEVEL to NOTICE.
@@ -3561,22 +3605,6 @@ Troubleshooting Qt 'configure' failures
 
 """
             )
-            # output_log = os.path.join(
-            #     builddir, "CMakeFiles", "CMakeOutput.log"
-            # )
-            # if os.path.exists(output_log):
-            #     print(80 * "*", file=sys.stderr)
-            #     print("Contents of CMakeOutput.log:", file=sys.stderr)
-            #     with open(output_log) as f:
-            #         print(f.read(), file=sys.stderr)
-
-            # error_log = os.path.join(builddir, "CMakeFiles", "CMakeError.log")  # noqa: E501
-            # if os.path.exists(error_log):
-            #     print(80 * "*", file=sys.stderr)
-            #     print("Contents of CMakeError.log:", file=sys.stderr)
-            #     with open(error_log) as f:
-            #         print(f.read(), file=sys.stderr)
-
             sys.exit(EXIT_FAILURE)
 
     # -------------------------------------------------------------------------