+++ docs/source/changelog.rst
@@ -3672,7 +3672,7 @@ Current C++/SQLite client, Python/SQLAlchemy server
   <edeq>`. (Database revision 0072.)
 
 - **New task:** :ref:`Internet Severity and Activities Addiction Questionnaire
-  (ISAAQ) <isaaq>`. (Database revision 0073.)
+  (ISAAQ) <isaaq10>`. (Database revision 0073.)
 
 - **New task:** :ref:`Internet Severity and Activities Addiction Questionnaire,
   Eating Disorders Appendix (ISAAQ-ED) <isaaqed>`. (Database revision 0074.)
@@ -3745,3 +3745,12 @@ Current C++/SQLite client, Python/SQLAlchemy server
 
 - Bug fix: Make BMI waist circumference optional again.
   https://github.com/ucam-department-of-psychiatry/camcops/issues/264
+
+- Update :ref:`Cambridge-Chicago Compulsivity Trait Scale (CHI-T) <chit>` to
+  remove q16 and add "Neither agree nor disagree" to the responses. This
+  means that the maximum total score is now 60 instead of 45.
+
+  (Database revision 0080.)
+
+- Remove ISAAQ task and replace with :ref:`ISAAQ-10 <isaaq10>`. (Database
+  revisions 0081-0083.)